Added files via upload
- Hotfix to prevent platform from tilting too far in any direction

Basically, as the servos reach their maximum rotation, they have to turn more in order to have the same effect (non-linearity in action). For some reason, sometimes a servo will try to switch from its highest angle to its lowest angle. This is a really hacky fix to prevent it from tilting too far. Ultimately this may be fine for simple tests, but I'd like to figure out why this is happening and eliminate it if possible.