(unstable) PID w/ neg. fbk from IMU -- ATMega
Different implementation of PID controller in SP code. 

This version compares the tilt seem by the IMU (placed on the moving platform) to the reference tilt (zero) and then adjusts the tilt of the platform by some amount. The roll and pitch are stored as global variables - these are what the PID controllers modify.

Unfortunately, this has some problems:
- Stability: platform *should* be adjusting towards center, but actually will tilt away at a speed proportional to its tilt (imagine a marble on an upside-down bowl)  
- Code size: for some reason this also runs at >90% of the arduino Uno's dynamic memory size and breaks at runtime. This is mitigated with the ATMega, which has 4X as much dynamic memory allocated.

I think if we can figure out a work-around for the error-accumulation, we'd be in good shape with this. Else, we still have the PID that we demo'd before.